Add prototype test for sign in & sign up

+++ internal/entites/user.go
@@ -10,9 +10,9 @@ const (
 )
 
 type User struct {
-	ID        int32        `db:"id"`
+	ID        uint64       `db:"id"`
 	Login     string       `db:"login"`
-	Inn       uint32       `db:"inn"`
+	Inn       uint64       `db:"inn"`
 	Name      string       `db:"name"`
 	Password  string       `db:"password"`
 	CreatedAt time.Time    `db:"created_at"`
